Create informal testing suite
This package needs some sanity check tests to make sure small changes
dont break big things. The `informal_tests` folder contains a
test script corresponding to each file in the `R/` folder.

A script loads the package using `devtools::load_all()` and grabs
sensitive information from a local config file. Tests are written
to cover each function in the package, and outputs can be dumped to a
local folder.

The config file should not be published to github because it contains
filepaths around our HPC, which should be kept private for security
in case this R package ever becomes public.