Make package public
This is a breaking change. We remove the default argument for the
`get_gft()` function to remove dependence on the Kwan lab file system.
Now, a file path to a GTF must be provided.

Additionally, `perform_edger()` can now receive for the `gtf` argument
 either a GTF dataframe
object or a file path to a GTF.